#36 - 예외핸들러 통합
