+++ src/main/java/com/artfolio/artfolio/global/error/ErrorHandlingController.java
@@ -1,9 +1,7 @@
 package com.artfolio.artfolio.global.error;
 
-import com.artfolio.artfolio.global.exception.AuctionAlreadyExistsException;
-import com.artfolio.artfolio.global.exception.AuctionAlreadyFinishedException;
-import com.artfolio.artfolio.global.exception.AuctionNotFoundException;
-import lombok.RequiredArgsConstructor;
+import com.artfolio.artfolio.global.exception.*;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.MethodArgumentNotValidException;
@@ -11,6 +9,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
+import org.springframework.web.reactive.function.client.WebClientResponseException;
 
 import java.util.List;
 
@@ -18,9 +17,51 @@ import static com.artfolio.artfolio.global.util.ErrorBuildFactory.*;
 
 @Slf4j
 @RestControllerAdvice
-@RequiredArgsConstructor
-public class AuctionErrorHandlingController {
-    /* validation 검증 실패시 발생하는 예외 핸들링 메서드 */
+public class ErrorHandlingController {
+    @ExceptionHandler(EntityNotFoundException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    protected ErrorResponse handleEntityNotFoundException() {
+        log.error("해당 엔티티를 찾을 수 없습니다.");
+        return buildError(ErrorCode.USER_NOT_FOUND);
+    }
+
+    @ExceptionHandler(UserNotFoundException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    protected ErrorResponse handleMemberNotFoundException() {
+        log.error("프론트로부터 잘못된 멤버 id가 전달되었습니다.");
+        return buildError(ErrorCode.USER_NOT_FOUND);
+    }
+
+    /* 존재하지 않는 예술품 번호가 넘어온 경우 */
+    @ExceptionHandler(ArtPieceNotFoundException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    protected ErrorResponse handleArtPieceNotFoundException(ArtPieceNotFoundException e) {
+        log.error("해당 예술품 번호가 존재하지 않습니다.");
+        log.error("art piece ID : " + e.getArtPieceId());
+        return buildError(ErrorCode.ARTPIECE_NOT_FOUND);
+    }
+
+    @ExceptionHandler(DeleteAuthorityException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    protected ErrorResponse handleMemberNotAuthorityException() {
+        log.error("해당 예술가가 등록한 경매가 아니여서 경매를 삭제할 수 없습니다.");
+        return buildError(ErrorCode.NO_DELETE_AUTHORITY);
+    }
+
+    @ExceptionHandler(DuplicateIdException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    protected ErrorResponse handleDuplicateIdException () {
+        log.error("이미 존재하는 회원입니다.");
+        return buildError(ErrorCode.DUPLICATE_ID);
+    }
+
+    @ExceptionHandler(WebClientResponseException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    protected ErrorResponse handleWebClientResponseException() {
+        log.error("OpenAI 크레딧 부족");
+        return buildError(ErrorCode.OPENAI_NOT_AVAILABLE);
+    }
+
     @ExceptionHandler(MethodArgumentNotValidException.class)
     @ResponseStatus(HttpStatus.BAD_REQUEST)
     protected ErrorResponse handleMethodArgumentNotValidException(MethodArgumentNotValidException e) {
